email-connector/requirements.txt
pypff-python>=20231205
libratom>=0.5.0
markdownify>=0.11.6
qdrant-client>=1.7.0
requests>=2.31.0
pymupdf>=1.23.0
python-docx>=1.1.0
